Update 02 including  renumbering Roadmap.rst
Renumbered the items: item 1 is now as below. I added the Panel App Gallery to your links with examples, for a total of 3 example links. 

The APP Gallery has Windturbines Explorer and VTK Slicer with brain MRI image near the top, so that seems appropriate. And always good to have links that refer from one Holoviz site to another!

1.  **New and improved workflow examples in biomedical and Earth sciences, especially utilizing machine learning**: Demonstrations for handling and visualization of various large, multidimensional datasets from data-intensive research domains, including on HPC systems. Update: Work is ongoing, but several such examples are now at https://projectpythia.org, https://github.com/holoviz-topics/neuro and https://panel.holoviz.org/gallery/index.html#app-gallery .